Add git http access using git-http-backend
This will be a major change. Primarily, it adds access to any
previously-known git repositories under the given repository directory
path. This is passed as the first argument on the command line.

All URLs containing '.git' are passed on to the git-http-backend CGI
executable. The executable is located by executing 'git --exec-path' on
the system command line, then appending '/git-http-backend'.