Add webui framework
The ShowPath() function should return a complete HTML webpage created
from the contents of a directory, the path of which is passed as the
argument. If the directory does not contain a '.git' file, then it
should be displayed as a directory listing. If it does, it should be
treated as a repository directory, and information should be given as to
how to clone it, as well as the log, and branches.