Added the html/css for log.
Now it shows the 10 most recent commits. That can be changed at some
point if we want it to change. I just didn't want to show like 300 commits
on some that have a shitton of commits. So yeah, maybe eventually we'll
change it to something else, but the most recent 10 is good for now.

Closes #14