Got specific commits for files working.
Still need to test the specific commits to make sure they're correct.
As of right now it looks like they're alright from a glance.
Also updated the .gitignore file to ignore .DS_STORE files (an OS X file).
Made a quick change to git.go to make the number of commits work.
Added an if statement to webui.go to figure out if it's a file or not.