Add install script
The install script will install Grove to '/usr/bin/' and its resources
to '/usr/share/grove/'. It has the option to skip the build process, and
also to build with any flags or compilers.

This commit will also add the "fmt" package to the imports, but other
dependencies already import this. This is so that Grove can report to
stdout, rather than stderr.