Stabilize grove.sh and adjust options
The script will now perform a more intelligent scan for source
directories, looking for `~/src` and `~/dev` before checking if the
current directory is a git repository. If so, it will serve that. If
not, it will fail.

The `DEV` variable has been renamed to `SRC`, and the default `LOG`
option has been set to `/dev/null`