Fixed an error with webui.go, and began to restyle Grove.
The error with webui.go was that every page contained ".."
When the root page should not contain ".." in the list and
Added "/" to go back to the root directory when not in the
Root directory.

Started to restyle Grove. Decided that there should be two
Different themes. A dark theme, and a light theme. Without
Specifying a flag, the light theme should be the default.

Ref #100