Reuse template rather than reread files
This is strongly encouraged in the little documentation for templates,
and on Stack Overflow. I added a list of template basenames (to prevent
any possible screwups with additional .html files being in
the *fRes/templates/ directory) and wrapper for creating the template
object.

Also, minor gofmts. @lukevers, you need emacs in your life.

Close #108